ccw-clientupdateinfo-memberRegis

+++ src/main/java/com/edu/neu/foodclient/entity/Client.java
@@ -13,7 +13,7 @@ import java.util.Date;
  */
 public class Client {
 	
-	private int clientid;
+	private Integer clientid;
 	private String username;
 	private String password;
 	private String wechat_id;
@@ -21,9 +21,9 @@ public class Client {
 	private String phone;
 	private String createdate;
 	private Member member;
-	private int type;
-	private int state;
-	private int scores;
+	private Integer type;
+	private Integer state;
+	private Double scores;
 
 	public Member getMember() {
 		return member;
@@ -50,83 +50,83 @@ public class Client {
 				'}';
 	}
 
-	public int getClientid() {
+	public Integer getClientid() {
 		return clientid;
 	}
-	
-	public void setClientid(int clientid) {
+
+	public void setClientid(Integer clientid) {
 		this.clientid = clientid;
 	}
-	
+
 	public String getUsername() {
 		return username;
 	}
-	
+
 	public void setUsername(String username) {
 		this.username = username;
 	}
-	
+
 	public String getPassword() {
 		return password;
 	}
-	
+
 	public void setPassword(String password) {
 		this.password = password;
 	}
-	
+
 	public String getWechat_id() {
 		return wechat_id;
 	}
-	
+
 	public void setWechat_id(String wechat_id) {
 		this.wechat_id = wechat_id;
 	}
-	
+
 	public String getIcon() {
 		return icon;
 	}
-	
+
 	public void setIcon(String icon) {
 		this.icon = icon;
 	}
-	
+
 	public String getPhone() {
 		return phone;
 	}
-	
+
 	public void setPhone(String phone) {
 		this.phone = phone;
 	}
-	
+
 	public String getCreatedate() {
 		return createdate;
 	}
-	
+
 	public void setCreatedate(String createdate) {
 		this.createdate = createdate;
 	}
-	
-	public int getType() {
+
+	public Integer getType() {
 		return type;
 	}
-	
-	public void setType(int type) {
+
+	public void setType(Integer type) {
 		this.type = type;
 	}
-	
-	public int getState() {
+
+	public Integer getState() {
 		return state;
 	}
-	
-	public void setState(int state) {
+
+	public void setState(Integer state) {
 		this.state = state;
 	}
-	
-	public int getScores() {
+
+	public Double getScores() {
 		return scores;
 	}
-	
-	public void setScores(int scores) {
+
+	public void setScores(Double scores) {
 		this.scores = scores;
 	}
 }
